requirements.txt
PyGithub
sentence-transformers
scikit-learn
numpy